requirements.txt
flask
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv
gunicorn
Flask-Migrate
config
requests
psycopg2
gunicorn